backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2
python-multipart==0.0.6
openai>=1.6.1
anthropic==0.7.7
chromadb==0.4.18
langchain==0.0.340
langchain-openai==0.0.2
langchain-text-splitters==0.0.1
pypdf2==3.0.1
python-docx==0.8.11
tiktoken>=0.5.2
pytest==7.4.3
pytest-asyncio==0.21.1
python-dotenv==1.0.0
aiofiles==23.2.1